Create ClassGeneratorBase and start moving common functionality there
