+++ python/podio_class_generator.py
@@ -12,12 +12,11 @@ from enum import IntEnum
 from collections.abc import Mapping
 from collections import defaultdict
 
-import jinja2
-
 from podio_schema_evolution import DataModelComparator  # dealing with cyclic imports
 from podio_schema_evolution import RenamedMember, root_filter, RootIoRule
 from podio_gen.podio_config_reader import PodioConfigReader
 from podio_gen.generator_utils import DataType, DefinitionError, DataModelJSONEncoder
+from podio_gen.generator_base import ClassGeneratorBase, write_file_if_changed
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
 TEMPLATE_DIR = os.path.join(THIS_DIR, 'templates')
@@ -69,24 +68,6 @@ def clang_format_file(content, name):
     return cfproc.communicate(input=content.encode())[0].decode()
 
 
-def write_file_if_changed(filename, content, force_write=False):
-  """Write the file contents only if it has changed or if the file does not exist
-  yet. Return whether the file has been written or not"""
-  try:
-    with open(filename, 'r', encoding='utf-8') as infile:
-      existing_content = infile.read()
-      changed = existing_content != content
-  except FileNotFoundError:
-    changed = True
-
-  if changed or force_write:
-    with open(filename, 'w', encoding='utf-8') as outfile:
-      outfile.write(content)
-    return True
-
-  return False
-
-
 class IncludeFrom(IntEnum):
   """Enum to signify if an include is needed and from where it should come"""
   NOWHERE = 0  # No include needed
@@ -94,19 +75,15 @@ class IncludeFrom(IntEnum):
   EXTERNAL = 2  # include from an upstream datamodel
 
 
-class ClassGenerator:
+class ClassGenerator(ClassGeneratorBase):
   """The entry point for reading a datamodel definition and generating the
   necessary source code from it."""
   # pylint: disable=too-many-arguments
   def __init__(self, yamlfile, install_dir, package_name, io_handlers, proglang, verbose, dryrun,
                upstream_edm, old_description, evolution_file):
-    self.install_dir = install_dir
-    self.package_name = package_name
+    super().__init__(yamlfile, install_dir, package_name, verbose, dryrun, upstream_edm)
     self.io_handlers = io_handlers
     self.proglang = proglang
-    self.verbose = verbose
-    self.dryrun = dryrun
-    self.yamlfile = yamlfile
     # schema evolution specific code
     self.old_yamlfile = old_description
     self.evolution_file = evolution_file
@@ -120,26 +97,6 @@ class ClassGenerator:
     self.root_schema_component_names = set()
     self.root_schema_datatype_names = set()
     self.root_schema_iorules = set()
-
-    try:
-      self.datamodel = PodioConfigReader.read(yamlfile, package_name, upstream_edm)
-    except DefinitionError as err:
-      print(f'Error while generating the datamodel: {err}')
-      sys.exit(1)
-
-    self.env = jinja2.Environment(loader=jinja2.FileSystemLoader(TEMPLATE_DIR),
-                                  keep_trailing_newline=True,
-                                  lstrip_blocks=True,
-                                  trim_blocks=True)
-
-    self.get_syntax = self.datamodel.options["getSyntax"]
-    self.incfolder = self.datamodel.options['includeSubfolder']
-    self.expose_pod_members = self.datamodel.options["exposePODMembers"]
-    self.upstream_edm = upstream_edm
-
-    self.formatter_func = None
-    self.generated_files = []
-    self.any_changes = False
 
   def _process_parent_module(self, datamodel):
     """Process parent module of julia that contains constructor definitions
@@ -231,75 +188,6 @@ have resolvable schema evolution incompatibilities:")
     for summaryline in text.splitlines():
       print(summaryline)
     print()
-
-  def _eval_template(self, template, data, old_schema_data=None):
-    """Fill the specified template"""
-    # merge the info of data and the old schema into a single dict
-    if old_schema_data:
-      data['OneToOneRelations_old'] = old_schema_data['OneToOneRelations']
-      data['OneToManyRelations_old'] = old_schema_data['OneToManyRelations']
-      data['VectorMembers_old'] = old_schema_data['VectorMembers']
-
-    return self.env.get_template(template).render(data)
-
-  def _write_file(self, name, content):
-    """Write the content to file. Dispatch to the correct directory depending on
-    whether it is a header or a .cc file."""
-    if name.endswith("h") or name.endswith("jl"):
-      fullname = os.path.join(self.install_dir, self.package_name, name)
-    else:
-      fullname = os.path.join(self.install_dir, "src", name)
-    if not self.dryrun:
-      self.generated_files.append(fullname)
-      if self.formatter_func:
-        content = self.formatter_func(content, fullname)
-
-      changed = write_file_if_changed(fullname, content)
-      self.any_changes = changed or self.any_changes
-
-  @staticmethod
-  def _get_filenames_templates(template_base, name):
-    """Get the list of output filenames and corresponding template names"""
-    # depending on which category is passed different naming conventions apply
-    # for the generated files. Additionally not all categories need source files.
-    # Listing the special cases here
-    def get_fn_format(tmpl):
-      """Get a format string for the filename"""
-      prefix = {'MutableObject': 'Mutable'}
-      postfix = {'Data': 'Data',
-                 'Obj': 'Obj',
-                 'SIOBlock': 'SIOBlock',
-                 'Collection': 'Collection',
-                 'CollectionData': 'CollectionData',
-                 'MutableStruct': 'Struct'
-                 }
-
-      return f'{prefix.get(tmpl, "")}{{name}}{postfix.get(tmpl, "")}.{{end}}'
-
-    endings = {
-        'Data': ('h',),
-        'PrintInfo': ('h',),
-        'MutableStruct': ('jl',),
-        'ParentModule': ('jl',),
-        }.get(template_base, ('h', 'cc'))
-
-    fn_templates = []
-    for ending in endings:
-      template_name = f'{template_base}.{ending}.jinja2'
-      filename = get_fn_format(template_base).format(name=name, end=ending)
-      fn_templates.append((filename, template_name))
-
-    return fn_templates
-
-  def _fill_templates(self, template_base, data, old_schema_data=None):
-    """Fill the template and write the results to file"""
-    # Update the passed data with some global things that are the same for all
-    # files
-    data['package_name'] = self.package_name
-    data['use_get_syntax'] = self.get_syntax
-    data['incfolder'] = self.incfolder
-    for filename, template in self._get_filenames_templates(template_base, data['class'].bare_type):
-      self._write_file(filename, self._eval_template(template, data, old_schema_data))
 
   def _process_component(self, name, component):
     """Process one component"""
